+++ CHANGELOG.rst
@@ -6,6 +6,8 @@ Changelog for Metview's Python interface
 ------------------
 - the Fieldset constructor can now take a list of paths to GRIB files
   - e.g. a = Fieldset(path=['/path1/to/data1.grib', 'relpath/data2.grib'])
+- the result of a call to mcont() etc can now be modified, e.g.
+  - c = mv.mcont() ; c["contour_line_colour"] = "green" ; mv.plot(data, c)
 
 1.7.2
 ------------------